requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
configparser==7.2.0
cssselect==1.3.0
distro==1.9.0
duckduckgo_search==8.0.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
fsspec==2025.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.1
lxml==5.4.0
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.2
mdurl==0.1.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.5
openai==1.77.0
packaging==25.0
pillow==11.2.1
primp==0.15.0
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==14.0.0
sgmllib3k==1.0.0
six==1.17.0
smolagents==1.14.0
sniffio==1.3.1
soupsieve==2.7
tinysegmenter==0.3
tldextract==5.3.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
